move validator input related stuff into its own module
